refactor: qdump.Redirect -> shared.Redirect
